Move record to primary class file
