+++ src/System.Device.Gpio/System/Runtime/CompilerServices/IsExternalInit.cs
@@ -1,0 +1,10 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+#pragma warning disable CS1591
+namespace System.Runtime.CompilerServices
+{
+    public class IsExternalInit
+    {
+    }
+}
